Prevent unhandled Exception when window is closing
When trying to send a PubSub message and in that moment the window is closed, `graphql-transport-electron` throws unhandled `Object has been destroyed` exception when calling event.sender.send. It's not possible to handle the exception in app code, also in many cases it's not possible to check whether webContents is destroyed or not, because there are a few event loop ticks between calling PubSub.publish and webContents.send. It occurs very frequently when PubSub is used intensively.